Convert GDC results to data.frame

GDC results are typically returned as an R list structure. This method converts that R list structure to a data.frame. Some columns in the resulting data.frame may remain lists, but there will be one list element for each row though that list element may contain multiple values.

as.data.frame.GDCResults(x, row.names, optional, ...)

Arguments

x
a GDCResults object to be converted to a data.frame.

row.names
not used here; row.names are calculated from the data to maintain data integrity

optional
not used here; just for matching method call for as.data.frame

...
not used here; just for matching method call for as.data.frame

Value

a data.frame, potentially with list columns still present.

Note

The data.frame that is returned is the cartesian product of sub-data.frames that come from the GDC. This may lead to more than one row per entity id. It is up to the user to determine how to make rows unique per entity, if that is desired.

Examples

expands = c("diagnoses","diagnoses.treatments","annotations",
            "demographic","exposures")
head(cases() %>% expand(expands) %>% results() %>% as.data.frame())

#>                                case_id               updated_datetime.x
#> 1 00b16434-0b3a-4194-8714-74dccbebe06c 2017-03-04T16:39:19.244769-06:00
#> 2 1431dd24-bd56-51e3-bf5d-160635a4946e 2017-03-04T16:39:19.244769-06:00
#> 3 2154d147-d266-5af8-bccd-8ccf817e0325 2017-03-04T16:39:19.244769-06:00
#> 4 3128d9bf-71fd-4edb-8d07-98f53ef0432e 2017-03-04T16:39:19.244769-06:00
#> 5 384400c1-98d6-4cc3-8724-de6711f6208b 2017-03-04T16:39:19.244769-06:00
#> 6 43cc3c60-f0a6-4b44-bd7b-8ec4da71991f 2017-03-04T16:39:19.244769-06:00
#>     submitter_id.x                                   id
#> 1     TCGA-EJ-8468 00b16434-0b3a-4194-8714-74dccbebe06c
#> 2 TARGET-30-PANUIF 1431dd24-bd56-51e3-bf5d-160635a4946e
#> 3 TARGET-30-PANDVF 2154d147-d266-5af8-bccd-8ccf817e0325
#> 4     TCGA-BP-5184 3128d9bf-71fd-4edb-8d07-98f53ef0432e
#> 5     TCGA-AA-3980 384400c1-98d6-4cc3-8724-de6711f6208b
#> 6     TCGA-2G-AAKG 43cc3c60-f0a6-4b44-bd7b-8ec4da71991f
#>                        disease_type created_datetime.x state.x   primary_site
#> 1           Prostate Adenocarcinoma                 NA    live       Prostate
#> 2                     Neuroblastoma                 NA    live Nervous System
#> 3                     Neuroblastoma                 NA    live Nervous System
#> 4 Kidney Renal Clear Cell Carcinoma                 NA    live         Kidney
#> 5              Colon Adenocarcinoma                 NA    live     Colorectal
#> 6       Testicular Germ Cell Tumors                 NA    live         Testis
#>   submitter_analyte_ids  analyte_ids diagnoses.classification_of_tumor
#> 1          TCGA-EJ-.... 27957f46....                      not reported
#> 2                                                                 <NA>
#> 3                                                                 <NA>
#> 4          TCGA-BP-.... 2a7a3764....                      not reported
#> 5          TCGA-AA-.... 83524c37....                      not reported
#> 6          TCGA-2G-.... 641cc611....                      not reported
#>   diagnoses.last_known_disease_status       diagnoses.updated_datetime
#> 1                        not reported 2017-03-04T16:44:35.784223-06:00
#> 2                                <NA> 2017-03-04T16:44:35.784223-06:00
#> 3                                <NA> 2017-03-04T16:44:35.784223-06:00
#> 4                        not reported 2017-03-04T16:44:35.784223-06:00
#> 5                        not reported 2017-03-04T16:44:35.784223-06:00
#> 6                        not reported 2017-03-04T16:44:35.784223-06:00
#>   diagnoses.primary_diagnosis diagnoses.submitter_id diagnoses.tumor_stage
#> 1                         c61 TCGA-EJ-8468_diagnosis          not reported
#> 2                not reported                   <NA>               stage 4
#> 3                not reported                   <NA>               stage 4
#> 4                       c64.9 TCGA-BP-5184_diagnosis               stage i
#> 5                       c18.9 TCGA-AA-3980_diagnosis               stage i
#> 6                      c62.90 TCGA-2G-AAKG_diagnosis               stage i
#>   diagnoses.age_at_diagnosis diagnoses.vital_status diagnoses.morphology
#> 1                      23371                  alive               8550/3
#> 2                       2234                  alive         not reported
#> 3                        687                  alive         not reported
#> 4                      19746                  alive               8310/3
#> 5                      32630                  alive               8140/3
#> 6                       8353                  alive               9070/3
#>   diagnoses.days_to_death diagnoses.days_to_last_known_disease_status
#> 1                      NA                                          NA
#> 2                      NA                                          NA
#> 3                      NA                                          NA
#> 4                      NA                                          NA
#> 5                      NA                                          NA
#> 6                      NA                                          NA
#>         diagnoses.created_datetime diagnoses.state diagnoses.days_to_recurrence
#> 1                             <NA>            live                           NA
#> 2 2016-04-25T16:38:51.018464-05:00            live                           NA
#> 3 2016-04-25T16:38:10.490777-05:00            live                           NA
#> 4                             <NA>            live                           NA
#> 5                             <NA>            live                           NA
#> 6                             <NA>            live                           NA
#>                 diagnoses.diagnosis_id diagnoses.tumor_grade
#> 1 c8a637ee-806c-5150-85ed-607970de34ac          not reported
#> 2 e48cfe76-39e9-52e2-9223-591d6037805b                  <NA>
#> 3 bf60ef4d-2f63-57ef-b0d3-9e91c99585b5                  <NA>
#> 4 042438f1-8e0f-5b54-926f-d8fbb5bb7285          not reported
#> 5 459930e0-5f0f-5abe-8d68-f0fa595f29b6          not reported
#> 6 db06405e-2dc7-508a-a431-2148705c0124          not reported
#>                                                                                                       diagnoses.treatments
#> 1 2017-03-04T16:43:59.646072-06:00, NA, NA, TCGA-EJ-8468_treatment, 95bc7afa-2674-5855-b946-b3723c1b8f57, live, NA, NA, NA
#> 2                                                                                                                     NULL
#> 3                                                                                                                     NULL
#> 4 2017-03-04T16:43:59.646072-06:00, NA, NA, TCGA-BP-5184_treatment, fa905b39-6605-5c01-9f0e-c657d1fabf74, live, NA, NA, NA
#> 5 2017-03-04T16:43:59.646072-06:00, NA, NA, TCGA-AA-3980_treatment, 697d562a-0c5f-5861-9409-fde97fc4003a, live, NA, NA, NA
#> 6 2017-03-04T16:43:59.646072-06:00, NA, NA, TCGA-2G-AAKG_treatment, d9868fb3-1bdb-5f8f-a04e-407725cb3889, live, NA, NA, NA
#>   diagnoses.tissue_or_organ_of_origin diagnoses.days_to_birth
#> 1                               c61.9                  -23371
#> 2                                <NA>                      NA
#> 3                                <NA>                      NA
#> 4                               c64.9                  -19746
#> 5                               c18.9                  -32630
#> 6                               c62.9                   -8353
#>   diagnoses.progression_or_recurrence diagnoses.prior_malignancy
#> 1                        not reported               not reported
#> 2                                <NA>                       <NA>
#> 3                                <NA>                       <NA>
#> 4                        not reported               not reported
#> 5                        not reported               not reported
#> 6                        not reported               not reported
#>   diagnoses.site_of_resection_or_biopsy diagnoses.days_to_last_follow_up
#> 1                                 c61.9                             2347
#> 2                          not reported                               NA
#> 3                          not reported                               NA
#> 4                                 c64.9                             1133
#> 5                                 c18.9                              822
#> 6                                 c62.9                             6598
#>     sample_ids  portion_ids submitter_portion_ids    slide_ids  aliquot_ids
#> 1 9e35383a.... 67ec88f5....          TCGA-EJ-.... a608635a.... 2e59dba2....
#> 2 9f8ef6cd....                                                 0d27f179....
#> 3                                                                          
#> 4 7be0587d.... 404f64b1....          TCGA-BP-.... 9cd1148b.... 5d1f5125....
#> 5 2c2efcd3.... a7f4e650....          TCGA-AA-.... 32b1615a.... 86469cd1....
#> 6 3a919db9.... ca387590....          TCGA-2G-.... 4ea5a324.... eed03778....
#>   submitter_aliquot_ids submitter_sample_ids exposures.cigarettes_per_day
#> 1          TCGA-EJ-....         TCGA-EJ-....                           NA
#> 2          TARGET-3....         TARGET-3....                           NA
#> 3                                                                      NA
#> 4          TCGA-BP-....         TCGA-BP-....                           NA
#> 5          TCGA-AA-....         TCGA-AA-....                           NA
#> 6          TCGA-2G-....         TCGA-2G-....                           NA
#>   exposures.weight       exposures.updated_datetime exposures.alcohol_history
#> 1               NA 2017-03-04T16:37:25.850486-06:00                      <NA>
#> 2               NA                             <NA>                      <NA>
#> 3               NA                             <NA>                      <NA>
#> 4               NA 2017-03-04T16:37:25.850486-06:00                      <NA>
#> 5               NA 2017-03-04T16:37:25.850486-06:00                      <NA>
#> 6               NA 2017-03-04T16:37:25.850486-06:00                      <NA>
#>   exposures.alcohol_intensity exposures.bmi exposures.years_smoked
#> 1                          NA            NA                     NA
#> 2                          NA            NA                     NA
#> 3                          NA            NA                     NA
#> 4                          NA            NA                     NA
#> 5                          NA            NA                     NA
#> 6                          NA            NA                     NA
#>   exposures.submitter_id exposures.created_datetime exposures.state
#> 1  TCGA-EJ-8468_exposure                         NA            live
#> 2                   <NA>                         NA            <NA>
#> 3                   <NA>                         NA            <NA>
#> 4  TCGA-BP-5184_exposure                         NA            live
#> 5  TCGA-AA-3980_exposure                         NA            live
#> 6  TCGA-2G-AAKG_exposure                         NA            live
#>                  exposures.exposure_id exposures.height submitter_slide_ids
#> 1 0f56213a-b63d-5e0f-840c-fc789725419b               NA        TCGA-EJ-....
#> 2                                 <NA>               NA                    
#> 3                                 <NA>               NA                    
#> 4 7c407073-c0e5-5466-bdc7-097f09458096               NA        TCGA-BP-....
#> 5 6b8e27c8-914e-5f75-a5ca-3e29226b075c               NA        TCGA-AA-....
#> 6 f383c3ae-cb40-5570-b100-615548317993               NA        TCGA-2G-....
#>                 updated_datetime.y               created_datetime.y gender
#> 1 2017-03-04T16:37:28.862150-06:00                             <NA>   male
#> 2 2017-03-04T16:37:28.862150-06:00 2016-04-25T16:38:51.018464-05:00 female
#> 3 2017-03-04T16:37:28.862150-06:00 2016-04-25T16:38:10.490777-05:00   male
#> 4 2017-03-04T16:37:28.862150-06:00                             <NA>   male
#> 5 2017-03-04T16:37:28.862150-06:00                             <NA> female
#> 6 2017-03-04T16:37:28.862150-06:00                             <NA>   male
#>   year_of_birth           submitter_id.y state.y         race
#> 1          1944 TCGA-EJ-8468_demographic    live not reported
#> 2            NA                     <NA>    live        white
#> 3            NA                     <NA>    live        white
#> 4          1953 TCGA-BP-5184_demographic    live        white
#> 5          1920 TCGA-AA-3980_demographic    live not reported
#> 6          1974 TCGA-2G-AAKG_demographic    live        white
#>                         demographic_id              ethnicity year_of_death
#> 1 b889907b-ecd1-521e-9247-7f3cbcbcea51           not reported            NA
#> 2 8f58f754-c23d-5e3f-808b-bc9338f75f4c not hispanic or latino            NA
#> 3 891be448-a045-569d-870c-9bdb1c2e5752 not hispanic or latino            NA
#> 4 a7ad2b63-3dd0-53b2-bb59-b5c6fd37d163           not reported            NA
#> 5 a3ccf839-9d7f-50c3-91c0-bd0e4441823d           not reported            NA
#> 6 125a69b4-cd09-5214-ac70-9193cf0cbfd3 not hispanic or latino            NA