Set facets for a

Set facets for a GDCQuery

Get facets for a GDCQuery

facet(x, facets)

get_facets(x)

# S3 method for GDCQuery
get_facets(x)

Arguments

x

a GDCQuery object

facets

a character vector of fields that will be used for forming aggregations (facets). Default is to set facets for all default fields. See default_fields for details

Examples

# create a new GDCQuery against the projects endpoint
gProj = projects()

# default facets are NULL
get_facets(gProj)

#> NULL

# set facets and save result
gProjFacet = facet(gProj)

# check facets
get_facets(gProjFacet)

#> [1] "dbgap_accession_number" "disease_type"           "name"                  
#> [4] "primary_site"           "project_id"             "released"              
#> [7] "state"                 

# and get a response, noting that
# the aggregations list member contains
# tibbles for each facet
str(response(gProjFacet,size=2),max.level=2)

#> List of 4
#>  $ results     :List of 2
#>   ..$ :List of 7
#>   ..$ :List of 7
#>  $ query       :List of 4
#>   ..$ fields : chr [1:7] "dbgap_accession_number" "disease_type" "name" "primary_site" ...
#>   ..$ filters: NULL
#>   ..$ facets : chr [1:7] "dbgap_accession_number" "disease_type" "name" "primary_site" ...
#>   ..$ token  : NULL
#>   ..- attr(*, "class")= chr [1:3] "gdc_projects" "GDCQuery" "list"
#>  $ pages       :List of 7
#>   ..$ count: int 2
#>   ..$ sort : chr ""
#>   ..$ from : int 1
#>   ..$ page : int 1
#>   ..$ total: int 39
#>   ..$ pages: int 20
#>   ..$ size : int 2
#>  $ aggregations:List of 7
#>   ..$ dbgap_accession_number:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	7 obs. of  2 variables:
#>   ..$ name                  :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	38 obs. of  2 variables:
#>   ..$ released              :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1 obs. of  2 variables:
#>   ..$ state                 :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1 obs. of  2 variables:
#>   ..$ primary_site          :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	29 obs. of  2 variables:
#>   ..$ project_id            :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	39 obs. of  2 variables:
#>   ..$ disease_type          :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	38 obs. of  2 variables: