return gdc-client executable path

This function is a convenience function to find and return the path to the GDC Data Transfer Tool executable assumed to be named 'gdc-client'. The assumption is that the appropriate version of the GDC Data Transfer Tool is a separate download available from https://gdc.cancer.gov/access-data/gdc-data-transfer-tool and as a backup from https://github.com/NCI-GDC/gdc-client. The following locations are checked:

gdc_client()

Value

character(1) the path to the gdc-client executable.

Details

Sys.which() to see if gdc-client is on the path

The current working directory

The file name specified in the environment variable GDC_CLIENT

Examples

# this cannot run without first
# downloading the GDC Data Transfer Tool
gdc_client = try(gdc_client(),silent=TRUE)