Get clinical information from GDC

The NCI GDC has a complex data model that allows various studies to supply numerous clinical and demographic data elements. However, across all projects that enter the GDC, there are similarities. This function returns four data.frames associated with case_ids from the GDC.

gdc_clinical(case_ids, include_list_cols = FALSE)

Arguments

case_ids
a character() vector of case_ids, typically from "cases" query.

include_list_cols
logical(1), whether to include list columns in the "main" data.frame. These list columns have values for aliquots, samples, etc. While these may be useful for some situations, they are generally not that useful as clinical annotations.

Value

A list of four data.frames:

main, representing basic case identification and metadata (update date, etc.)

diagnoses

esposures

demographic

Details

Note that these data.frames can, in general, have different numbers of rows (or even no rows at all). If one wishes to combine to produce a single data.frame, using the approach of left joining to the "main" data.frame will yield a useful combined data.frame. We do not do that directly given the potential for 1:many relationships. It is up to the user to determine what the best approach is for any given dataset.

Examples

case_ids = cases() %>% results(size=10) %>% ids()
clinical_data = gdc_clinical(case_ids)

# overview of clinical results
class(clinical_data)

#> [1] "GDCClinicalList" "list"           

names(clinical_data)

#> [1] "demographic" "diagnoses"   "exposures"   "main"       

sapply(clinical_data, class)

#>      demographic  diagnoses    exposures    main        
#> [1,] "tbl_df"     "tbl_df"     "tbl_df"     "tbl_df"    
#> [2,] "tbl"        "tbl"        "tbl"        "tbl"       
#> [3,] "data.frame" "data.frame" "data.frame" "data.frame"

sapply(clinical_data, nrow)

#> demographic   diagnoses   exposures        main 
#>          10          10           0          10 

# available data
head(clinical_data$main)

#> # A tibble: 6 x 8
#>   updated_datetime submitter_id id    disease_type created_datetime state
#>   <chr>            <chr>        <chr> <chr>        <chr>            <chr>
#> 1 2018-08-07T05:0… AD3978       4033… Adenomas an… 2017-06-01T09:3… rele…
#> 2 2018-08-07T05:0… AD14467      c8a4… Not Reported 2017-06-01T08:5… rele…
#> 3 2018-08-07T05:0… AD17381      0b55… Squamous Ce… 2017-06-01T09:1… rele…
#> 4 2018-08-07T05:0… AD11339      bcd7… Adenomas an… 2017-06-01T08:4… rele…
#> 5 2018-08-07T05:0… AD3547       4304… Adenomas an… 2017-06-01T09:3… rele…
#> 6 2018-08-07T05:0… AD15350      0946… Complex Mix… 2017-06-01T09:0… rele…
#> # ... with 2 more variables: case_id <chr>, primary_site <chr>

head(clinical_data$demographic)

#> # A tibble: 6 x 9
#>   updated_datetime created_datetime gender submitter_id state race 
#>   <chr>            <chr>            <chr>  <chr>        <chr> <chr>
#> 1 2018-08-07T05:0… 2017-06-19T11:2… female AD3978_demo… rele… not …
#> 2 2018-08-07T05:0… 2017-06-19T11:3… male   AD14467_dem… rele… not …
#> 3 2018-08-07T05:0… 2017-06-19T11:4… male   AD17381_dem… rele… not …
#> 4 2018-08-07T05:0… 2017-06-19T11:2… male   AD11339_dem… rele… not …
#> 5 2018-08-07T05:0… 2017-06-19T11:2… male   AD3547_demo… rele… not …
#> 6 2018-08-07T05:0… 2017-06-19T11:4… female AD15350_dem… rele… not …
#> # ... with 3 more variables: demographic_id <chr>, ethnicity <chr>,
#> #   case_id <chr>

head(clinical_data$diagnoses)

#> # A tibble: 6 x 20
#>   case_id progression_or_… classification_… last_known_dise… tumor_grade
#>   <chr>   <chr>            <chr>            <chr>            <chr>      
#> 1 403380… not reported     primary          not reported     not report…
#> 2 c8a4f0… not reported     metastasis       not reported     not report…
#> 3 0b5511… not reported     Unknown          not reported     not report…
#> 4 bcd702… not reported     primary          not reported     not report…
#> 5 4304fe… not reported     primary          not reported     not report…
#> 6 0946a9… not reported     metastasis       not reported     not report…
#> # ... with 15 more variables: tissue_or_organ_of_origin <chr>,
#> #   created_datetime <chr>, updated_datetime <chr>, primary_diagnosis <chr>,
#> #   submitter_id <chr>, site_of_resection_or_biopsy <chr>, tumor_stage <chr>,
#> #   state <chr>, age_at_diagnosis <int>, vital_status <chr>, morphology <chr>,
#> #   days_to_last_known_disease_status <lgl>, days_to_last_follow_up <lgl>,
#> #   days_to_recurrence <lgl>, diagnosis_id <chr>

head(clinical_data$exposures)

#> # A tibble: 0 x 1
#> # ... with 1 variable: case_id <chr>