Download GDC files

Download one or more files from GDC. Files are downloaded using the UUID and renamed to the file name on the remote system. By default, neither the uuid nor the file name on the remote system can exist.

gdcdata(uuids, use_cached = TRUE, progress = interactive(),
  token = NULL, access_method = "api", transfer_args = character())

Arguments

uuids
character() of GDC file UUIDs.

use_cached
logical(1) default TRUE indicating that, if found in the cache, the file will not be downloaded again. If FALSE, all supplied uuids will be re-downloaded.

progress
logical(1) default TRUE in interactive sessions, FALSE otherwise indicating whether a progress par should be produced for each file download.

token
(optional) character(1) security token allowing access to restricted data. See https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/.

access_method
character(1), either 'api' or 'client'. See details.

transfer_args
character(1), additional arguments to pass to the gdc-client command line. See gdc_client and transfer_help for details.

Value

a named vector with file uuids as the names and paths as the value

Details

This function is appropriate for one or several files; for large downloads use manifest to create a manifest for and the GDC Data Transfer Tool.

When access_method is "api", the GDC "data" endpoint is the transfer mechanism used. The alternative access_method, "client", will utilize the gdc-client transfer tool, which must be downloaded separately and available. See gdc_client for details on specifying the location of the gdc-client executable.

See also

Examples

# get some example file uuids
uuids <- files() %>%
    filter(~ access == 'open' & file_size < 100000) %>%
    results(size = 3) %>%
    ids()

# and get the data, placing it into the gdc_cache() directory
fpaths <- gdcdata(uuids, use_cached=TRUE)

fpaths

#>                                                                                                                               6f88847b-5643-4f79-96f0-79bec5e1ed33 
#> "/Users/sdavis2/Library/Caches/GenomicDataCommons/6f88847b-5643-4f79-96f0-79bec5e1ed33/RIELS_p_TCGA_268_269_N_GenomeWideSNP_6_E12_1306390.nocnv_grch38.seg.v2.txt" 
#>                                                                                                                               b8ab150a-a778-47eb-9c7a-cb6e2225d6c8 
#>   "/Users/sdavis2/Library/Caches/GenomicDataCommons/b8ab150a-a778-47eb-9c7a-cb6e2225d6c8/REEDY_p_TCGAb65_SNP_N_GenomeWideSNP_6_D11_697596.nocnv_grch38.seg.v2.txt" 
#>                                                                                                                               a12bbc40-fe3f-4baf-9d22-029b01c04a34 
#>   "/Users/sdavis2/Library/Caches/GenomicDataCommons/a12bbc40-fe3f-4baf-9d22-029b01c04a34/AMISS_p_TCGAb24_SNP_N_GenomeWideSNP_6_H06_542442.nocnv_grch38.seg.v2.txt"