Create NCI GDC filters for limiting GDC query results

Create filters using standard R formatting

make_filter(expr, available_fields = NULL, toJSON = TRUE)

Arguments

expr

an R expression

available_fields

A character vector of field names that, if specified, can be used unquoted in creating filter expressions

toJSON

boolean, return JSON (if true) or just a list

Value

a JSON data object (if toJSON==TRUE, the default) can be used as the filter parameter in an NCI GDC search or other query.

Details

Searching the NCI GDC allows for complex filtering based on logical operations and simple comparisons. This function facilitates writing such filter expressions in R-like syntax with R code evaluation.

This function can be used in several different ways.

If used with available_fields, "bare" fields that are named in the available_fields character vector can be used in the filter expression without quotes.

If toJSON==TRUE, then the appropriate JSON for using in a query filter (filter = make_filter(....,)) is returned.

If toJSON==FALSE, then a list is returned that could be manipulated further before converrsion to JSON for use in a query. Note that the appropriate toJSON() will likely need to include “auto_unbox=TRUE”.

Examples

make_filter("diagnoses.age_at_diagnosis" <= 10*365)

#> {"op":"<=","content":{"field":"diagnoses.age_at_diagnosis","value":3650}} 

make_filter("project.primary_site" %in% c('blood','brain'))

#> {"op":"in","content":{"field":"project.primary_site","value":["blood","brain"]}} 

fe = make_filter("project.primary_site" %in% c('blood','brain')
                 & "diagnosis.age_at_diagnosis" <= 10*365)
fe

#> {"op":"and","content":[{"op":"in","content":{"field":"project.primary_site","value":["blood","brain"]}},{"op":"<=","content":{"field":"diagnosis.age_at_diagnosis","value":3650}}]} 

fe = make_filter("project.primary_site" %in% c('blood','brain')
                 & "diagnosis.age_at_diagnosis" <= 10*365,toJSON=FALSE)
fe

#> $op
#> [x] "and"
#> 
#> $content
#> $content[[1]]
#> $content[[1]]$op
#> [x] "in"
#> 
#> $content[[1]]$content
#> $content[[1]]$content$field
#> [1] "project.primary_site"
#> 
#> $content[[1]]$content$value
#> [1] "blood" "brain"
#> 
#> 
#> 
#> $content[[2]]
#> $content[[2]]$op
#> [x] "<="
#> 
#> $content[[2]]$content
#> $content[[2]]$content$field
#> [1] "diagnosis.age_at_diagnosis"
#> 
#> $content[[2]]$content$value
#> [1] 3650
#> 
#> 
#> 
#>