Prepare GDC manifest file for bulk download

The manifest function/method creates a manifest of files to be downloaded using the GDC Data Transfer Tool. There are methods for creating manifest data frames from GDCQuery objects that contain file information ("cases" and "files" queries).

manifest(x, from = 0, size = count(x), ...)

# S3 method for gdc_files
manifest(x, from = 0, size = count(x), ...)

# S3 method for GDCfilesResponse
manifest(x, from = 0, size = count(x), ...)

# S3 method for GDCcasesResponse
manifest(x, from = 0, size = count(x), ...)

Arguments

x
An GDCQuery object of subclass "gdc_files" or "gdc_cases".

from
Record number from which to start when returning the manifest.

size
The total number of records to return. Default will return the usually desirable full set of records.

...
passed to PUT.

Value

A tibble, also of type "gdc_manifest", with five columns:

id

filename

md5

size

state

Methods (by class)

gdc_files:

GDCfilesResponse:

GDCcasesResponse:

Examples

gFiles = files()
shortManifest = gFiles %>% manifest(size=10)
head(shortManifest,n=3)

#> # A tibble: 3 x 5
#>                                     id
#>                                  <chr>
#> 1 c5c4b4a3-3224-4a72-a883-c99c7747e47b
#> 2 dd029237-d470-4b58-9cf0-11753fa60972
#> 3 3fe677f6-8329-447c-b999-5e70582624aa
#> # ... with 4 more variables: filename <chr>, md5 <chr>, size <int>, state <chr>