Query GDC for available endpoint fields

Query GDC for available endpoint fields

mapping(endpoint)

Arguments

endpoint

character(1) corresponding to endpoints for which users may specify additional or alternative fields. Endpoints include “projects”, “cases”, “files”, and “annotations”.

Value

A data frame describing the field (field name), full (full data model name), type (data type), and four additional columns describing the "set" to which the fields belong--“default”, “expand”, “multi”, and “nested”.

Examples

map <- mapping("projects")
head(map)

#>                                            description doc_type
#> 1                                                      projects
#> 2                                                      projects
#> 3                                                      projects
#> 4                                                      projects
#> 5 The dbgap accession number provided for the program. projects
#> 6                      Full name/title of the program. projects
#>                            field                                    full   type
#> 1         dbgap_accession_number         projects.dbgap_accession_number string
#> 2                   disease_type                   projects.disease_type     id
#> 3                           name                           projects.name     id
#> 4                   primary_site                   projects.primary_site     id
#> 5 program.dbgap_accession_number projects.program.dbgap_accession_number string
#> 6                   program.name                   projects.program.name string
#>   defaults expand multi nested
#> 1     TRUE  FALSE FALSE  FALSE
#> 2     TRUE  FALSE  TRUE  FALSE
#> 3     TRUE  FALSE  TRUE  FALSE
#> 4     TRUE  FALSE  TRUE  FALSE
#> 5    FALSE  FALSE FALSE  FALSE
#> 6    FALSE  FALSE FALSE  FALSE

# get only the "default" fields
subset(map,defaults)

#>    description doc_type                  field                            full
#> 1              projects dbgap_accession_number projects.dbgap_accession_number
#> 2              projects           disease_type           projects.disease_type
#> 3              projects                   name                   projects.name
#> 4              projects           primary_site           projects.primary_site
#> 8              projects             project_id             projects.project_id
#> 9              projects               released               projects.released
#> 10             projects                  state                  projects.state
#>      type defaults expand multi nested
#> 1  string     TRUE  FALSE FALSE  FALSE
#> 2      id     TRUE  FALSE  TRUE  FALSE
#> 3      id     TRUE  FALSE  TRUE  FALSE
#> 4      id     TRUE  FALSE  TRUE  FALSE
#> 8      id     TRUE  FALSE  TRUE  FALSE
#> 9  string     TRUE  FALSE FALSE  FALSE
#> 10 string     TRUE  FALSE FALSE  FALSE

# And get just the text names of the "default" fields
subset(map,defaults)$field

#> [1] "dbgap_accession_number" "disease_type"           "name"                  
#> [4] "primary_site"           "project_id"             "released"              
#> [7] "state"