Start a query of GDC metadata

The basis for all functionality in this package starts with constructing a query in R. The GDCQuery object contains the filters, facets, and other parameters that define the returned results. A token is required for accessing certain datasets.

query(entity, filters = NULL, facets = NULL, legacy = FALSE,
  expand = NULL, fields = default_fields(entity))

cases(...)

files(...)

projects(...)

annotations(...)

Arguments

entity
character vector of 'cases','files','annotations', or 'projects'

filters
a filter list, typically created using make_filter, or added to an existing GDCQuery object using filter.

facets
a character vector of

legacy
logical(1) whether to use the "legacy" archive or not. See https://docs.gdc.cancer.gov/Data_Portal/Users_Guide/Legacy_Archive/ and https://gdc-portal.nci.nih.gov/legacy-archive/search/f for details.

expand
a character vector of "expands" to include in returned data

fields
a character vector of fields to return

...
passed through to query

Value

An S3 object, the GDCQuery object. This is a list with the following members.

filters

facets

fields

expand

archive

token

Functions

cases: convenience contructor for a GDCQuery for cases

files: convenience contructor for a GDCQuery for cases

projects: convenience contructor for a GDCQuery for cases

annotations: convenience contructor for a GDCQuery for annotations

Examples

qcases = query('cases')
# equivalent to:
qcases = cases()