Bulk data download

The GDC maintains a special tool, https://docs.gdc.cancer.gov/Data_Transfer_Tool/Users_Guide/Getting_Started/, that enables high-performance, potentially parallel, and resumable downloads. The Data Transfer Tool is an external program that requires separate download. #' @param gdc_client character(1) name or path to gdc-client executable. The executable that is used is found through the gdc_client. See gdc_client for details on how to set the executable path.

transfer(uuids, args = character(), token = NULL, overwrite = FALSE)

transfer_help()

Arguments

uuids
character() vector of GDC file UUIDs

args
character() vector specifying command-line arguments to be passed to gdc-client. See transfer_help for possible values. The arguments --manifest, --dir, and --token-file are determined by manifest, destination_dir, and token, respectively, and should NOT be provided as elements of args.

token
character(1) containing security token allowing access to restricted data. See https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/. Note that the GDC transfer tool requires a file for data transfer. Therefore, this token will be written to a temporary file (with appropriate permissions set).

overwrite
logical(1) default FALSE indicating whether existing files with identical name should be over-written.

Value

character(1) directory path to which the files were downloaded.

Functions

transfer_help:

Examples

file_manifest = files() %>% filter(~ access == "open") %>% manifest(size=10)
manifest_file = tempfile()
write.table(file_manifest,file=manifest_file,col.names=TRUE,row.names=FALSE,quote=FALSE)
destination <- transfer(manifest_file)

#> Error: '{"data": {"hits": [], "pagination": {"count": 0, "sort": "", "from": 0, "pages": 0, "total": 0, "page": 0, "size": 0}}, "warnings": {}}' does not exist in current working directory ('/Users/sdavis2/Documents/git/GenomicDataCommons/docs/reference').

dir(destination)

#> Error in dir(destination): object 'destination' not found

# and with authenication
destination <- transfer(manifest_file,token=gdc_token)

#> Error: '{"data": {"hits": [], "pagination": {"count": 0, "sort": "", "from": 0, "pages": 0, "total": 0, "page": 0, "size": 0}}, "warnings": {}}' does not exist in current working directory ('/Users/sdavis2/Documents/git/GenomicDataCommons/docs/reference').