Manipulating GDCQuery filters

Manipulating GDCQuery filters

The filter is simply a safe accessor for the filter element in GDCQuery objects.

The get_filter is simply a safe accessor for the filter element in GDCQuery objects.

filter(x, expr)

# S3 method for GDCQuery
filter(x, expr)

get_filter(x)

# S3 method for GDCQuery
get_filter(x)

Arguments

x

the object on which to set the filter list member

expr

a filter expression, where bare names (without quotes) are allowed if they are available fields associated with the x

Examples

# make a GDCQuery object to start
#
# Projects
#
pQuery = projects()

#> Error in curl::curl_fetch_memory(url, handle = handle): SSL connect error

# check for the default fields
# so that we can use one of them to build a filter
default_fields(pQuery)

#> Error in default_fields(pQuery): object 'pQuery' not found

pQuery = filter(pQuery,~ project_id == 'TCGA-LUAC')

#> Error in filter(pQuery, ~project_id == "TCGA-LUAC"): object 'pQuery' not found

get_filter(pQuery)

#> Error in get_filter(pQuery): object 'pQuery' not found

#
# Files
#
fQuery = files()

#> Error in curl::curl_fetch_memory(url, handle = handle): SSL connect error

default_fields(fQuery)

#> Error in default_fields(fQuery): object 'fQuery' not found

fQuery = filter(fQuery,~ data_format == 'VCF')

#> Error in filter(fQuery, ~data_format == "VCF"): object 'fQuery' not found

get_filter(fQuery)

#> Error in get_filter(fQuery): object 'fQuery' not found

fQuery = filter(fQuery,~ data_format == 'VCF' & experimental_strategy == 'WXS' & type == 'simple_somatic_mutation')

#> Error in filter(fQuery, ~data_format == "VCF" & experimental_strategy ==     "WXS" & type == "simple_somatic_mutation"): object 'fQuery' not found

# Use str() to get a cleaner picture
str(get_filter(fQuery))

#> Error in get_filter(fQuery): object 'fQuery' not found