Get clinical information from GDC

The NCI GDC has a complex data model that allows various studies to supply numerous clinical and demographic data elements. However, across all projects that enter the GDC, there are similarities. This function returns four data.frames associated with case_ids from the GDC.

gdc_clinical(case_ids, include_list_cols = FALSE)

Arguments

case_ids
a character() vector of case_ids, typically from "cases" query.

include_list_cols
logical(1), whether to include list columns in the "main" data.frame. These list columns have values for aliquots, samples, etc. While these may be useful for some situations, they are generally not that useful as clinical annotations.

Value

A list of four data.frames:

main, representing basic case identification and metadata (update date, etc.)

diagnoses

esposures

demographic

Details

Note that these data.frames can, in general, have different numbers of rows (or even no rows at all). If one wishes to combine to produce a single data.frame, using the approach of left joining to the "main" data.frame will yield a useful combined data.frame. We do not do that directly given the potential for 1:many relationships. It is up to the user to determine what the best approach is for any given dataset.

Examples

case_ids = cases() %>% results(size=10) %>% ids()
clinical_data = gdc_clinical(case_ids)

# overview of clinical results
class(clinical_data)

#> [1] "list"

names(clinical_data)

#> [1] "demographic" "diagnoses"   "exposures"   "main"       

sapply(clinical_data, class)

#>  demographic    diagnoses    exposures         main 
#> "data.frame" "data.frame" "data.frame" "data.frame" 

sapply(clinical_data, nrow)

#> demographic   diagnoses   exposures        main 
#>          10          10           2          10 

# available data
head(clinical_data$main)

#>                   updated_datetime     submitter_id
#> 1 2018-08-07T05:08:25.415044+00:00          AD11218
#> 2 2018-08-07T05:08:25.415044+00:00           AD8623
#> 3 2018-05-22T01:01:29.874989-05:00     TCGA-13-0727
#> 4 2018-08-07T05:08:25.415044+00:00           AD1214
#> 5 2018-05-21T18:18:57.804483-05:00     TCGA-W5-AA2X
#> 6 2017-12-15T21:14:49.455236-06:00 TARGET-50-PADXJZ
#>                                     id                      disease_type
#> 1 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf      Ductal and Lobular Neoplasms
#> 2 4060f92b-ab42-4233-8609-6e9baa11631d                           Gliomas
#> 3 90f4b65c-cfd4-4066-a5b9-842885c172e2 Ovarian Serous Cystadenocarcinoma
#> 4 a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10      Ductal and Lobular Neoplasms
#> 5 6e534dd3-81ac-4575-ab1e-02c21e10916d                Cholangiocarcinoma
#> 6 dfd8aaad-cf31-5260-8882-b471ddbcc8fd             High-Risk Wilms Tumor
#>                   created_datetime    state
#> 1 2017-06-01T08:35:27.577245-05:00 released
#> 2 2017-06-01T09:06:24.248479-05:00 released
#> 3                             <NA>     live
#> 4 2017-06-01T09:13:35.077867-05:00 released
#> 5                             <NA>     live
#> 6                             <NA>     live
#>                                case_id
#> 1 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf
#> 2 4060f92b-ab42-4233-8609-6e9baa11631d
#> 3 90f4b65c-cfd4-4066-a5b9-842885c172e2
#> 4 a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10
#> 5 6e534dd3-81ac-4575-ab1e-02c21e10916d
#> 6 dfd8aaad-cf31-5260-8882-b471ddbcc8fd
#>                                                             primary_site
#> 1                                                                 Breast
#> 2 Spinal cord, cranial nerves, and other parts of central nervous system
#> 3                                                                  Ovary
#> 4                                                                 Breast
#> 5                                                              Bile Duct
#> 6                                                                 Kidney

head(clinical_data$demographic)

#>                                                      updated_datetime
#> 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf 2018-08-07T05:08:25.415044+00:00
#> 4060f92b-ab42-4233-8609-6e9baa11631d 2018-08-07T05:08:25.415044+00:00
#> 90f4b65c-cfd4-4066-a5b9-842885c172e2 2018-03-09T08:04:27.510418-06:00
#> a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10 2018-08-07T05:08:25.415044+00:00
#> 6e534dd3-81ac-4575-ab1e-02c21e10916d 2018-03-06T09:56:05.842739-06:00
#> dfd8aaad-cf31-5260-8882-b471ddbcc8fd 2017-12-11T08:30:41.387981-06:00
#>                                                      created_datetime gender
#> 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf 2017-06-19T11:23:58.646681-05:00 female
#> 4060f92b-ab42-4233-8609-6e9baa11631d 2017-06-19T11:44:35.217289-05:00   male
#> 90f4b65c-cfd4-4066-a5b9-842885c172e2                             <NA> female
#> a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10 2017-06-19T11:48:22.177979-05:00 female
#> 6e534dd3-81ac-4575-ab1e-02c21e10916d                             <NA>   male
#> dfd8aaad-cf31-5260-8882-b471ddbcc8fd 2017-02-24T20:54:45.808091-06:00   male
#>                                                      submitter_id    state
#> 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf          AD11218_demographic released
#> 4060f92b-ab42-4233-8609-6e9baa11631d           AD8623_demographic released
#> 90f4b65c-cfd4-4066-a5b9-842885c172e2     TCGA-13-0727_demographic     live
#> a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10           AD1214_demographic released
#> 6e534dd3-81ac-4575-ab1e-02c21e10916d     TCGA-W5-AA2X_demographic     live
#> dfd8aaad-cf31-5260-8882-b471ddbcc8fd TARGET-50-PADXJZ_demographic     live
#>                                              race
#> 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf not reported
#> 4060f92b-ab42-4233-8609-6e9baa11631d not reported
#> 90f4b65c-cfd4-4066-a5b9-842885c172e2        white
#> a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10 not reported
#> 6e534dd3-81ac-4575-ab1e-02c21e10916d        white
#> dfd8aaad-cf31-5260-8882-b471ddbcc8fd        white
#>                                                            demographic_id
#> 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf bdb16ed7-17c8-427b-8bf6-0c0102a7db83
#> 4060f92b-ab42-4233-8609-6e9baa11631d d03eda99-4387-408d-88fa-1bc5ce303024
#> 90f4b65c-cfd4-4066-a5b9-842885c172e2 c607c1f9-1240-5281-9ad1-0481bf7a2684
#> a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10 c60c3464-416d-4294-bd2d-def66d2874ab
#> 6e534dd3-81ac-4575-ab1e-02c21e10916d 16df8d49-14b4-57b9-bb70-f288356797ce
#> dfd8aaad-cf31-5260-8882-b471ddbcc8fd 3d277971-599f-5fe5-94fe-a534c80ce743
#>                                                   ethnicity year_of_birth
#> 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf           not reported            NA
#> 4060f92b-ab42-4233-8609-6e9baa11631d           not reported            NA
#> 90f4b65c-cfd4-4066-a5b9-842885c172e2 not hispanic or latino          1932
#> a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10           not reported            NA
#> 6e534dd3-81ac-4575-ab1e-02c21e10916d not hispanic or latino          1943
#> dfd8aaad-cf31-5260-8882-b471ddbcc8fd not hispanic or latino            NA
#>                                      year_of_death
#> 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf            NA
#> 4060f92b-ab42-4233-8609-6e9baa11631d            NA
#> 90f4b65c-cfd4-4066-a5b9-842885c172e2          2004
#> a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10            NA
#> 6e534dd3-81ac-4575-ab1e-02c21e10916d          2010
#> dfd8aaad-cf31-5260-8882-b471ddbcc8fd            NA
#>                                                                   case_id
#> 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf
#> 4060f92b-ab42-4233-8609-6e9baa11631d 4060f92b-ab42-4233-8609-6e9baa11631d
#> 90f4b65c-cfd4-4066-a5b9-842885c172e2 90f4b65c-cfd4-4066-a5b9-842885c172e2
#> a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10 a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10
#> 6e534dd3-81ac-4575-ab1e-02c21e10916d 6e534dd3-81ac-4575-ab1e-02c21e10916d
#> dfd8aaad-cf31-5260-8882-b471ddbcc8fd dfd8aaad-cf31-5260-8882-b471ddbcc8fd

head(clinical_data$diagnoses)

#>                                case_id progression_or_recurrence
#> 1 1f95a9f1-0327-4c89-a6fe-69809d0d8ebf              not reported
#> 2 4060f92b-ab42-4233-8609-6e9baa11631d              not reported
#> 3 90f4b65c-cfd4-4066-a5b9-842885c172e2              not reported
#> 4 a4ec8b0b-0b53-43e6-90b4-e88a8a3cbc10              not reported
#> 5 6e534dd3-81ac-4575-ab1e-02c21e10916d              not reported
#> 6 dfd8aaad-cf31-5260-8882-b471ddbcc8fd              not reported
#>   classification_of_tumor last_known_disease_status  tumor_grade
#> 1                 primary              not reported not reported
#> 2              metastasis              not reported not reported
#> 3            not reported              not reported not reported
#> 4                 primary              not reported not reported
#> 5            not reported              not reported not reported
#> 6                    <NA>              not reported Not Reported
#>   tissue_or_organ_of_origin                 created_datetime
#> 1               Breast, NOS 2017-06-16T15:41:54.034178-05:00
#> 2       Nervous system, NOS 2017-06-19T09:23:29.064224-05:00
#> 3                     C56.9                             <NA>
#> 4               Breast, NOS 2017-06-19T09:29:52.147488-05:00
#> 5                     C24.0                             <NA>
#> 6              Not Reported 2017-02-24T20:54:45.808091-06:00
#>                   updated_datetime                primary_diagnosis
#> 1 2018-08-07T05:08:25.415044+00:00 Infiltrating duct carcinoma, NOS
#> 2 2018-08-07T05:08:25.415044+00:00           Oligodendroglioma, NOS
#> 3 2018-03-09T16:42:15.753583-06:00                            C56.9
#> 4 2018-08-07T05:08:25.415044+00:00 Infiltrating duct carcinoma, NOS
#> 5 2018-03-08T19:30:04.867636-06:00                            C24.0
#> 6 2018-01-19T13:39:21.801433-06:00                     not reported
#>                 submitter_id site_of_resection_or_biopsy  tumor_stage    state
#> 1          AD11218_diagnosis                 Breast, NOS not reported released
#> 2           AD8623_diagnosis                  Brain, NOS not reported released
#> 3     TCGA-13-0727_diagnosis                       C56.9 not reported     live
#> 4           AD1214_diagnosis                 Breast, NOS not reported released
#> 5     TCGA-W5-AA2X_diagnosis                       C24.0    stage ivb     live
#> 6 TARGET-50-PADXJZ_diagnosis                        C649          iii     live
#>   age_at_diagnosis vital_status morphology days_to_last_known_disease_status
#> 1            24657 not reported     8500/3                                NA
#> 2            10112 not reported     9450/3                                NA
#> 3            26185         dead     8441/3                                NA
#> 4            21113 not reported     8500/3                                NA
#> 5            24797         dead     8160/3                                NA
#> 6              838        alive     8960/3                                NA
#>   days_to_last_follow_up days_to_recurrence
#> 1                     NA                 NA
#> 2                     NA                 NA
#> 3                    462                 NA
#> 4                     NA                 NA
#> 5                     NA                 NA
#> 6                   4311                 NA
#>                           diagnosis_id days_to_death days_to_birth
#> 1 a9bdec94-8c84-466b-9d32-17239313c0cd            NA            NA
#> 2 e94992aa-e877-43cc-9c57-414348b264f5            NA            NA
#> 3 b8106532-5977-55ab-8efb-28955ed4a91a           462        -26185
#> 4 705df2cb-7ad0-42da-bc69-8526bb6cfa08            NA            NA
#> 5 1f12809f-8b6d-57f7-b45b-91392aa7b2cd           271        -24797
#> 6 ca2ff673-3653-5f13-af7e-7ec8e1e0cb91            NA            NA
#>   prior_malignancy
#> 1             <NA>
#> 2             <NA>
#> 3     not reported
#> 4             <NA>
#> 5     not reported
#> 6             <NA>

head(clinical_data$exposures)

#>                                case_id cigarettes_per_day weight
#> 1 90f4b65c-cfd4-4066-a5b9-842885c172e2                 NA     NA
#> 2 6e534dd3-81ac-4575-ab1e-02c21e10916d                 NA     88
#>                   updated_datetime alcohol_history alcohol_intensity      bmi
#> 1 2018-03-09T08:04:27.510418-06:00              NA                NA       NA
#> 2 2018-03-06T09:56:05.842739-06:00              NA                NA 31.55366
#>   years_smoked          submitter_id created_datetime state
#> 1           NA TCGA-13-0727_exposure               NA  live
#> 2           NA TCGA-W5-AA2X_exposure               NA  live
#>                            exposure_id height
#> 1 f820881b-3240-54d4-83e2-9fb8a687da8d     NA
#> 2 247910d0-5a86-552d-b096-22b59ba24032    167