Get clinical information from GDC

The NCI GDC has a complex data model that allows various studies to supply numerous clinical and demographic data elements. However, across all projects that enter the GDC, there are similarities. This function returns four data.frames associated with case_ids from the GDC.

Usage

gdc_clinical(case_ids, include_list_cols = FALSE)

Arguments

case_ids

a character() vector of case_ids, typically from "cases" query.

include_list_cols

logical(1), whether to include list columns in the "main" data.frame. These list columns have values for aliquots, samples, etc. While these may be useful for some situations, they are generally not that useful as clinical annotations.

Value

A list of four data.frames:

main, representing basic case identification and metadata (update date, etc.)

diagnoses

esposures

demographic

Details

Note that these data.frames can, in general, have different numbers of rows (or even no rows at all). If one wishes to combine to produce a single data.frame, using the approach of left joining to the "main" data.frame will yield a useful combined data.frame. We do not do that directly given the potential for 1:many relationships. It is up to the user to determine what the best approach is for any given dataset.

Examples

case_ids = cases() %>% results(size=10) %>% ids()
clinical_data = gdc_clinical(case_ids)

# overview of clinical results
class(clinical_data)
#> [1] "GDCClinicalList" "list"           
names(clinical_data)
#> [1] "demographic" "diagnoses"   "exposures"   "main"       
sapply(clinical_data, class)
#>      demographic  diagnoses    exposures    main        
#> [1,] "tbl_df"     "tbl_df"     "tbl_df"     "tbl_df"    
#> [2,] "tbl"        "tbl"        "tbl"        "tbl"       
#> [3,] "data.frame" "data.frame" "data.frame" "data.frame"
sapply(clinical_data, nrow)
#> demographic   diagnoses   exposures        main 
#>          10          14          10          10 

# available data
head(clinical_data$main)
#> # A tibble: 6 × 13
#>   id         lost_to_followup days_to_lost_to… created_datetime updated_datetime
#>   <chr>      <chr>            <lgl>            <chr>            <chr>           
#> 1 4829dd8c-… NA               NA               2020-07-31T09:2… 2021-07-12T12:2…
#> 2 d420e653-… NA               NA               2019-10-14T10:4… 2021-03-03T15:1…
#> 3 bfe15f44-… NA               NA               2019-08-14T15:1… 2020-10-20T19:2…
#> 4 8b3b1f24-… NA               NA               2018-10-02T15:5… 2019-09-20T15:1…
#> 5 fa30fc7f-… NA               NA               2020-05-18T13:2… 2021-07-12T12:2…
#> 6 47322ea3-… NA               NA               2019-10-14T10:4… 2021-07-12T12:2…
#> # … with 8 more variables: case_id <chr>, state <chr>, disease_type <chr>,
#> #   submitter_id <chr>, days_to_consent <lgl>, consent_type <lgl>,
#> #   primary_site <chr>, index_date <chr>
head(clinical_data$demographic)
#> # A tibble: 6 × 22
#>   cause_of_death race  gender ethnicity   vital_status age_at_index submitter_id
#>   <chr>          <chr> <chr>  <chr>       <chr>        <lgl>        <chr>       
#> 1 Cancer Related white female Unknown     Dead         NA           HCM-BROD-00…
#> 2 Cancer Related white female not hispan… Dead         NA           HCM-BROD-02…
#> 3 NA             white male   not hispan… Alive        NA           HCM-CSHL-00…
#> 4 NA             white female not hispan… Alive        NA           HCM-CSHL-00…
#> 5 Cancer Related white male   Unknown     Dead         NA           HCM-BROD-00…
#> 6 NA             white male   not hispan… Alive        NA           HCM-BROD-03…
#> # … with 15 more variables: days_to_birth <int>, created_datetime <chr>,
#> #   year_of_birth <int>, cause_of_death_source <lgl>, premature_at_birth <lgl>,
#> #   weeks_gestation_at_birth <lgl>, demographic_id <chr>,
#> #   updated_datetime <chr>, age_is_obfuscated <lgl>,
#> #   occupation_duration_years <lgl>, days_to_death <int>, state <chr>,
#> #   year_of_death <lgl>, country_of_residence_at_enrollment <lgl>,
#> #   case_id <chr>
head(clinical_data$diagnoses)
#> # A tibble: 6 × 101
#>   case_id gleason_pattern… ajcc_pathologic… ann_arbor_clini… created_datetime   
#>   <chr>   <lgl>            <chr>            <lgl>            <dttm>             
#> 1 4829dd… NA               NA               NA               2020-07-31 14:25:24
#> 2 d420e6… NA               NA               NA               2020-02-27 19:00:21
#> 3 bfe15f… NA               Stage I          NA               2020-02-27 18:37:46
#> 4 8b3b1f… NA               Stage I          NA               2019-05-15 18:02:25
#> 5 fa30fc… NA               Stage IV         NA               2020-06-15 17:57:49
#> 6 fa30fc… NA               NA               NA               2020-06-15 17:57:49
#> # … with 96 more variables: tissue_or_organ_of_origin <chr>,
#> #   age_at_diagnosis <int>, cog_liver_stage <lgl>,
#> #   child_pugh_classification <lgl>, peritoneal_fluid_cytological_status <lgl>,
#> #   metastasis_at_diagnosis_site <chr>, state <chr>,
#> #   pregnant_at_diagnosis <lgl>, cog_rhabdomyosarcoma_risk_group <lgl>,
#> #   primary_gleason_grade <lgl>, margins_involved_site <lgl>, morphology <chr>,
#> #   medulloblastoma_molecular_classification <lgl>, …
head(clinical_data$exposures)
#> # A tibble: 6 × 33
#>   case_id    alcohol_days_pe… alcohol_drinks_… smokeless_tobac… created_datetime
#>   <chr>                 <int>            <int> <lgl>            <chr>           
#> 1 4829dd8c-…               NA               NA NA               2020-07-31T09:2…
#> 2 d420e653-…               NA               NA NA               2020-02-27T13:0…
#> 3 bfe15f44-…               NA               NA NA               2020-02-27T12:3…
#> 4 8b3b1f24-…               NA               NA NA               2019-05-15T13:0…
#> 5 fa30fc7f-…               NA                0 NA               2020-06-15T12:5…
#> 6 47322ea3-…               NA               NA NA               2020-09-17T16:1…
#> # … with 28 more variables: pack_years_smoked <lgl>,
#> #   secondhand_smoke_as_child <lgl>, asbestos_exposure <chr>,
#> #   cigarettes_per_day <lgl>, age_at_onset <lgl>, tobacco_smoking_status <chr>,
#> #   updated_datetime <chr>, exposure_id <chr>, marijuana_use_per_week <lgl>,
#> #   tobacco_smoking_onset_year <lgl>,
#> #   respirable_crystalline_silica_exposure <lgl>, exposure_duration <lgl>,
#> #   years_smoked <lgl>, state <chr>, type_of_smoke_exposure <lgl>, …