Get clinical information from GDC

The NCI GDC has a complex data model that allows various studies to supply numerous clinical and demographic data elements. However, across all projects that enter the GDC, there are similarities. This function returns four data.frames associated with case_ids from the GDC.

Usage

gdc_clinical(case_ids, include_list_cols = FALSE)

Arguments

case_ids

a character() vector of case_ids, typically from "cases" query.

include_list_cols

logical(1), whether to include list columns in the "main" data.frame. These list columns have values for aliquots, samples, etc. While these may be useful for some situations, they are generally not that useful as clinical annotations.

Value

A list of four data.frames:

main, representing basic case identification and metadata (update date, etc.)

diagnoses

esposures

demographic

Details

Note that these data.frames can, in general, have different numbers of rows (or even no rows at all). If one wishes to combine to produce a single data.frame, using the approach of left joining to the "main" data.frame will yield a useful combined data.frame. We do not do that directly given the potential for 1:many relationships. It is up to the user to determine what the best approach is for any given dataset.

Examples

case_ids = cases() %>% results(size=10) %>% ids()
clinical_data = gdc_clinical(case_ids)

# overview of clinical results
class(clinical_data)
#> [1] "GDCClinicalList" "list"           
names(clinical_data)
#> [1] "demographic" "diagnoses"   "exposures"   "main"       
sapply(clinical_data, class)
#>      demographic  diagnoses    exposures    main        
#> [1,] "tbl_df"     "tbl_df"     "tbl_df"     "tbl_df"    
#> [2,] "tbl"        "tbl"        "tbl"        "tbl"       
#> [3,] "data.frame" "data.frame" "data.frame" "data.frame"
sapply(clinical_data, nrow)
#> demographic   diagnoses   exposures        main 
#>          10          10          10          10 

# available data
head(clinical_data$main)
#> # A tibble: 6 × 11
#>   id     lost_…¹ disea…² days_…³ submi…⁴ creat…⁵ prima…⁶ updat…⁷ case_id index…⁸
#>   <chr>  <lgl>   <chr>   <lgl>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 2525b… NA      Ductal… NA      C3N-03… 2019-0… Pancre… 2021-0… 2525bf… Diagno…
#> 2 12650… NA      Gliomas NA      C3N-01… 2018-0… Brain   2020-0… 126507… Diagno…
#> 3 c43ac… NA      Squamo… NA      C3N-03… 2019-0… Other … 2020-1… c43ac4… Diagno…
#> 4 a59a9… NA      Squamo… NA      C3N-02… 2019-0… Other … 2020-1… a59a90… Diagno…
#> 5 59122… NA      Adenom… NA      C3L-03… 2019-0… Bronch… 2022-0… 59122a… Diagno…
#> 6 4447a… NA      Gliomas NA      C3L-03… 2019-0… Brain   2020-0… 4447a9… Diagno…
#> # … with 1 more variable: state <chr>, and abbreviated variable names
#> #   ¹​lost_to_followup, ²​disease_type, ³​days_to_lost_to_followup, ⁴​submitter_id,
#> #   ⁵​created_datetime, ⁶​primary_site, ⁷​updated_datetime, ⁸​index_date
#> # ℹ Use `colnames()` to see all variable names
head(clinical_data$demographic)
#> # A tibble: 6 × 18
#>   cause_o…¹ race  gender ethni…² vital…³ age_a…⁴ submi…⁵ days_…⁶ creat…⁷ year_…⁸
#>   <chr>     <chr> <chr>  <chr>   <chr>   <lgl>   <chr>     <int> <chr>     <int>
#> 1 NA        not … female not re… Alive   NA      C3N-03…  -19956 2019-0…    1963
#> 2 Cancer R… not … female not re… Dead    NA      C3N-01…  -26312 2018-0…    1945
#> 3 NA        not … female not re… Alive   NA      C3N-03…  -25635 2019-0…    1947
#> 4 Cancer R… not … male   not re… Dead    NA      C3N-02…  -16652 2019-0…    1971
#> 5 NA        white male   not hi… Alive   NA      C3L-03…  -23384 2019-0…    1954
#> 6 Cancer R… white female not hi… Dead    NA      C3L-03…  -29326 2019-0…    1937
#> # … with 8 more variables: premature_at_birth <lgl>,
#> #   weeks_gestation_at_birth <lgl>, demographic_id <chr>,
#> #   updated_datetime <chr>, days_to_death <int>, state <chr>,
#> #   year_of_death <int>, case_id <chr>, and abbreviated variable names
#> #   ¹​cause_of_death, ²​ethnicity, ³​vital_status, ⁴​age_at_index, ⁵​submitter_id,
#> #   ⁶​days_to_birth, ⁷​created_datetime, ⁸​year_of_birth
#> # ℹ Use `colnames()` to see all variable names
head(clinical_data$diagnoses)
#> # A tibble: 6 × 82
#>   case_id    irs_s…¹ iss_s…² ajcc_…³ ann_a…⁴ created_datetime    ennek…⁵ inrg_…⁶
#>   <chr>      <lgl>   <lgl>   <chr>   <lgl>   <dttm>              <lgl>   <lgl>  
#> 1 2525bfef-… NA      NA      Stage … NA      2019-07-22 11:40:02 NA      NA     
#> 2 126507c3-… NA      NA      Not Re… NA      2018-12-03 18:05:16 NA      NA     
#> 3 c43ac461-… NA      NA      Stage … NA      2019-03-14 15:37:34 NA      NA     
#> 4 a59a90d9-… NA      NA      Not Re… NA      2019-03-14 15:37:34 NA      NA     
#> 5 59122a43-… NA      NA      Not Re… NA      2019-07-22 11:40:02 NA      NA     
#> 6 4447a969-… NA      NA      Not Re… NA      2019-05-07 12:41:33 NA      NA     
#> # … with 74 more variables: enneking_msts_metastasis <lgl>,
#> #   tissue_or_organ_of_origin <chr>, age_at_diagnosis <int>,
#> #   esophageal_columnar_dysplasia_degree <lgl>, cog_liver_stage <lgl>,
#> #   child_pugh_classification <lgl>, metastasis_at_diagnosis_site <lgl>,
#> #   state <chr>, prior_treatment <lgl>, cog_rhabdomyosarcoma_risk_group <lgl>,
#> #   days_to_last_known_disease_status <dbl>, primary_gleason_grade <lgl>,
#> #   inpc_grade <lgl>, ajcc_pathologic_t <chr>, morphology <chr>, …
#> # ℹ Use `colnames()` to see all variable names
head(clinical_data$exposures)
#> # A tibble: 6 × 25
#>   case_id        alcoh…¹ type_…² smoki…³ submi…⁴ type_…⁵ alcoh…⁶ envir…⁷ radon…⁸
#>   <chr>          <lgl>   <lgl>   <lgl>   <chr>   <lgl>   <lgl>   <lgl>   <lgl>  
#> 1 2525bfef-6962… NA      NA      NA      C3N-03… NA      NA      NA      NA     
#> 2 126507c3-c0d7… NA      NA      NA      C3N-01… NA      NA      NA      NA     
#> 3 c43ac461-9f03… NA      NA      NA      C3N-03… NA      NA      NA      NA     
#> 4 a59a90d9-f1b0… NA      NA      NA      C3N-02… NA      NA      NA      NA     
#> 5 59122a43-606a… NA      NA      NA      C3L-03… NA      NA      NA      NA     
#> 6 4447a969-e5c8… NA      NA      NA      C3L-03… NA      NA      NA      NA     
#> # … with 16 more variables: created_datetime <chr>, alcohol_intensity <chr>,
#> #   pack_years_smoked <dbl>, asbestos_exposure <lgl>, cigarettes_per_day <dbl>,
#> #   tobacco_smoking_quit_year <int>, tobacco_smoking_status <chr>,
#> #   alcohol_history <chr>, updated_datetime <chr>, exposure_id <chr>,
#> #   tobacco_smoking_onset_year <int>,
#> #   respirable_crystalline_silica_exposure <lgl>, years_smoked <dbl>,
#> #   state <chr>, time_between_waking_and_first_smoke <lgl>, …
#> # ℹ Use `colnames()` to see all variable names