Download GDC files

Download one or more files from GDC. Files are downloaded using the UUID and renamed to the file name on the remote system. By default, neither the uuid nor the file name on the remote system can exist.

Usage

gdcdata(
  uuids,
  use_cached = TRUE,
  progress = interactive(),
  token = NULL,
  access_method = "api",
  transfer_args = character(),
  ...
)

Arguments

uuids

character() of GDC file UUIDs.

use_cached

logical(1) default TRUE indicating that, if found in the cache, the file will not be downloaded again. If FALSE, all supplied uuids will be re-downloaded.

progress

logical(1) default TRUE in interactive sessions, FALSE otherwise indicating whether a progress par should be produced for each file download.

token

(optional) character(1) security token allowing access to restricted data. See https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/.

access_method

character(1), either 'api' or 'client'. See details.

transfer_args

character(1), additional arguments to pass to the gdc-client command line. See gdc_client and transfer_help for details.

...

further arguments passed to files, particulary useful when requesting legacy=TRUE

Value

a named vector with file uuids as the names and paths as the value

Details

This function is appropriate for one or several files; for large downloads use manifest to create a manifest for and the GDC Data Transfer Tool.

When access_method is "api", the GDC "data" endpoint is the transfer mechanism used. The alternative access_method, "client", will utilize the gdc-client transfer tool, which must be downloaded separately and available. See gdc_client for details on specifying the location of the gdc-client executable.

See also

Examples

# get some example file uuids
uuids <- files() %>%
    filter(~ access == 'open' & file_size < 100000) %>%
    results(size = 3) %>%
    ids()

# and get the data, placing it into the gdc_cache() directory
gdcdata(uuids, use_cached=TRUE)
#>                                                                                                              e4f53f7a-7bf4-4908-bbe7-4dee159e04b2 
#>        "~/.cache/GenomicDataCommons/e4f53f7a-7bf4-4908-bbe7-4dee159e04b2/a50a7ba0-c479-410f-a338-9ab179975875.wxs.aliquot_ensemble_masked.maf.gz" 
#>                                                                                                              77f11c00-5e1c-4258-a5be-fd367459fadc 
#> "~/.cache/GenomicDataCommons/77f11c00-5e1c-4258-a5be-fd367459fadc/TCGA-BRCA.c6e7c59d-708f-497e-b6e6-ed91ba294943.ascat2.allelic_specific.seg.txt" 
#>                                                                                                              78026596-7c57-403d-85a3-ca244cbb9117 
#>                                   "~/.cache/GenomicDataCommons/78026596-7c57-403d-85a3-ca244cbb9117/nationwidechildrens.org_ssf.TCGA-E9-A1NE.xml" 

# legacy data
exon <- files(legacy = TRUE) %>%
    filter( ~ cases.project.project_id == "TCGA-COAD" &
        data_category == "Gene expression" &
        data_type == "Exon quantification") %>%
    results(size = 1) %>% ids()

gdcdata(exon, legacy = TRUE)
#>                                                                                                               13f62ad4-b55d-4091-9cea-a4dce58f9943 
#> "~/.cache/GenomicDataCommons/13f62ad4-b55d-4091-9cea-a4dce58f9943/unc.edu.85eb9996-655c-44f8-abbf-fbc332cf75cf.1756113.bt.exon_quantification.txt"