Find matching field names

This utility function allows quick text-based search of available fields for using grep

grep_fields(entity, pattern, ..., value = TRUE)

Arguments

entity
one of "files", "cases", "annotations", "projects" against which to gather available fields for matching

pattern
A regular expression that will be used in a call to grep

...
passed on to grep

value
logical(1) whether to return values as opposed to indices (passed along to grep)

Value

character() vector of field names matching pattern

Examples

grep_fields('files','analysis')

#>  [1] "analysis.analysis_id"                                                     
#>  [2] "analysis.analysis_type"                                                   
#>  [3] "analysis.created_datetime"                                                
#>  [4] "analysis.input_files.access"                                              
#>  [5] "analysis.input_files.created_datetime"                                    
#>  [6] "analysis.input_files.data_category"                                       
#>  [7] "analysis.input_files.data_format"                                         
#>  [8] "analysis.input_files.data_type"                                           
#>  [9] "analysis.input_files.error_type"                                          
#> [10] "analysis.input_files.experimental_strategy"                               
#> [11] "analysis.input_files.file_id"                                             
#> [12] "analysis.input_files.file_name"                                           
#> [13] "analysis.input_files.file_size"                                           
#> [14] "analysis.input_files.file_state"                                          
#> [15] "analysis.input_files.md5sum"                                              
#> [16] "analysis.input_files.platform"                                            
#> [17] "analysis.input_files.revision"                                            
#> [18] "analysis.input_files.state"                                               
#> [19] "analysis.input_files.state_comment"                                       
#> [20] "analysis.input_files.submitter_id"                                        
#> [21] "analysis.input_files.updated_datetime"                                    
#> [22] "analysis.metadata.read_groups.adapter_name"                               
#> [23] "analysis.metadata.read_groups.adapter_sequence"                           
#> [24] "analysis.metadata.read_groups.base_caller_name"                           
#> [25] "analysis.metadata.read_groups.base_caller_version"                        
#> [26] "analysis.metadata.read_groups.created_datetime"                           
#> [27] "analysis.metadata.read_groups.experiment_name"                            
#> [28] "analysis.metadata.read_groups.flow_cell_barcode"                          
#> [29] "analysis.metadata.read_groups.includes_spike_ins"                         
#> [30] "analysis.metadata.read_groups.instrument_model"                           
#> [31] "analysis.metadata.read_groups.is_paired_end"                              
#> [32] "analysis.metadata.read_groups.library_name"                               
#> [33] "analysis.metadata.read_groups.library_preparation_kit_catalog_number"     
#> [34] "analysis.metadata.read_groups.library_preparation_kit_name"               
#> [35] "analysis.metadata.read_groups.library_preparation_kit_vendor"             
#> [36] "analysis.metadata.read_groups.library_preparation_kit_version"            
#> [37] "analysis.metadata.read_groups.library_selection"                          
#> [38] "analysis.metadata.read_groups.library_strand"                             
#> [39] "analysis.metadata.read_groups.library_strategy"                           
#> [40] "analysis.metadata.read_groups.platform"                                   
#> [41] "analysis.metadata.read_groups.read_group_id"                              
#> [42] "analysis.metadata.read_groups.read_group_name"                            
#> [43] "analysis.metadata.read_groups.read_group_qcs.adapter_content"             
#> [44] "analysis.metadata.read_groups.read_group_qcs.basic_statistics"            
#> [45] "analysis.metadata.read_groups.read_group_qcs.created_datetime"            
#> [46] "analysis.metadata.read_groups.read_group_qcs.encoding"                    
#> [47] "analysis.metadata.read_groups.read_group_qcs.fastq_name"                  
#> [48] "analysis.metadata.read_groups.read_group_qcs.kmer_content"                
#> [49] "analysis.metadata.read_groups.read_group_qcs.overrepresented_sequences"   
#> [50] "analysis.metadata.read_groups.read_group_qcs.per_base_n_content"          
#> [51] "analysis.metadata.read_groups.read_group_qcs.per_base_sequence_content"   
#> [52] "analysis.metadata.read_groups.read_group_qcs.per_base_sequence_quality"   
#> [53] "analysis.metadata.read_groups.read_group_qcs.per_sequence_gc_content"     
#> [54] "analysis.metadata.read_groups.read_group_qcs.per_sequence_quality_score"  
#> [55] "analysis.metadata.read_groups.read_group_qcs.per_tile_sequence_quality"   
#> [56] "analysis.metadata.read_groups.read_group_qcs.percent_gc_content"          
#> [57] "analysis.metadata.read_groups.read_group_qcs.read_group_qc_id"            
#> [58] "analysis.metadata.read_groups.read_group_qcs.sequence_duplication_levels" 
#> [59] "analysis.metadata.read_groups.read_group_qcs.sequence_length_distribution"
#> [60] "analysis.metadata.read_groups.read_group_qcs.state"                       
#> [61] "analysis.metadata.read_groups.read_group_qcs.submitter_id"                
#> [62] "analysis.metadata.read_groups.read_group_qcs.total_sequences"             
#> [63] "analysis.metadata.read_groups.read_group_qcs.updated_datetime"            
#> [64] "analysis.metadata.read_groups.read_group_qcs.workflow_end_datetime"       
#> [65] "analysis.metadata.read_groups.read_group_qcs.workflow_link"               
#> [66] "analysis.metadata.read_groups.read_group_qcs.workflow_start_datetime"     
#> [67] "analysis.metadata.read_groups.read_group_qcs.workflow_type"               
#> [68] "analysis.metadata.read_groups.read_group_qcs.workflow_version"            
#> [69] "analysis.metadata.read_groups.read_length"                                
#> [70] "analysis.metadata.read_groups.RIN"                                        
#> [71] "analysis.metadata.read_groups.sequencing_center"                          
#> [72] "analysis.metadata.read_groups.sequencing_date"                            
#> [73] "analysis.metadata.read_groups.size_selection_range"                       
#> [74] "analysis.metadata.read_groups.spike_ins_concentration"                    
#> [75] "analysis.metadata.read_groups.spike_ins_fasta"                            
#> [76] "analysis.metadata.read_groups.state"                                      
#> [77] "analysis.metadata.read_groups.submitter_id"                               
#> [78] "analysis.metadata.read_groups.target_capture_kit_catalog_number"          
#> [79] "analysis.metadata.read_groups.target_capture_kit_name"                    
#> [80] "analysis.metadata.read_groups.target_capture_kit_target_region"           
#> [81] "analysis.metadata.read_groups.target_capture_kit_vendor"                  
#> [82] "analysis.metadata.read_groups.target_capture_kit_version"                 
#> [83] "analysis.metadata.read_groups.to_trim_adapter_sequence"                   
#> [84] "analysis.metadata.read_groups.updated_datetime"                           
#> [85] "analysis.state"                                                           
#> [86] "analysis.submitter_id"                                                    
#> [87] "analysis.updated_datetime"                                                
#> [88] "analysis.workflow_end_datetime"                                           
#> [89] "analysis.workflow_link"                                                   
#> [90] "analysis.workflow_start_datetime"                                         
#> [91] "analysis.workflow_type"                                                   
#> [92] "analysis.workflow_version"                                                
#> [93] "downstream_analyses.analysis_id"                                          
#> [94] "downstream_analyses.analysis_type"