Find matching field names

This utility function allows quick text-based search of available fields for using grep

grep_fields(entity, pattern, ..., value = TRUE)

Arguments

entity
one of "files", "cases", "annotations", "projects" against which to gather available fields for matching

pattern
A regular expression that will be used in a call to grep

...
passed on to grep

value
logical(1) whether to return values as opposed to indices (passed along to grep)

Value

character() vector of field names matching pattern

Examples

grep_fields('files','analysis')

#>   [1] "analysis.analysis_id"                                                     
#>   [2] "analysis.analysis_type"                                                   
#>   [3] "analysis.created_datetime"                                                
#>   [4] "analysis.input_files.access"                                              
#>   [5] "analysis.input_files.average_base_quality"                                
#>   [6] "analysis.input_files.average_insert_size"                                 
#>   [7] "analysis.input_files.average_read_length"                                 
#>   [8] "analysis.input_files.created_datetime"                                    
#>   [9] "analysis.input_files.data_category"                                       
#>  [10] "analysis.input_files.data_format"                                         
#>  [11] "analysis.input_files.data_type"                                           
#>  [12] "analysis.input_files.error_type"                                          
#>  [13] "analysis.input_files.experimental_strategy"                               
#>  [14] "analysis.input_files.file_id"                                             
#>  [15] "analysis.input_files.file_name"                                           
#>  [16] "analysis.input_files.file_size"                                           
#>  [17] "analysis.input_files.imaging_date"                                        
#>  [18] "analysis.input_files.magnification"                                       
#>  [19] "analysis.input_files.md5sum"                                              
#>  [20] "analysis.input_files.mean_coverage"                                       
#>  [21] "analysis.input_files.pairs_on_diff_chr"                                   
#>  [22] "analysis.input_files.platform"                                            
#>  [23] "analysis.input_files.proportion_base_mismatch"                            
#>  [24] "analysis.input_files.proportion_coverage_10X"                             
#>  [25] "analysis.input_files.proportion_coverage_30X"                             
#>  [26] "analysis.input_files.proportion_reads_duplicated"                         
#>  [27] "analysis.input_files.proportion_reads_mapped"                             
#>  [28] "analysis.input_files.proportion_targets_no_coverage"                      
#>  [29] "analysis.input_files.read_pair_number"                                    
#>  [30] "analysis.input_files.revision"                                            
#>  [31] "analysis.input_files.state"                                               
#>  [32] "analysis.input_files.state_comment"                                       
#>  [33] "analysis.input_files.submitter_id"                                        
#>  [34] "analysis.input_files.total_reads"                                         
#>  [35] "analysis.input_files.updated_datetime"                                    
#>  [36] "analysis.metadata.read_groups.adapter_name"                               
#>  [37] "analysis.metadata.read_groups.adapter_sequence"                           
#>  [38] "analysis.metadata.read_groups.base_caller_name"                           
#>  [39] "analysis.metadata.read_groups.base_caller_version"                        
#>  [40] "analysis.metadata.read_groups.created_datetime"                           
#>  [41] "analysis.metadata.read_groups.days_to_sequencing"                         
#>  [42] "analysis.metadata.read_groups.experiment_name"                            
#>  [43] "analysis.metadata.read_groups.flow_cell_barcode"                          
#>  [44] "analysis.metadata.read_groups.fragment_maximum_length"                    
#>  [45] "analysis.metadata.read_groups.fragment_mean_length"                       
#>  [46] "analysis.metadata.read_groups.fragment_minimum_length"                    
#>  [47] "analysis.metadata.read_groups.fragment_standard_deviation_length"         
#>  [48] "analysis.metadata.read_groups.includes_spike_ins"                         
#>  [49] "analysis.metadata.read_groups.instrument_model"                           
#>  [50] "analysis.metadata.read_groups.is_paired_end"                              
#>  [51] "analysis.metadata.read_groups.lane_number"                                
#>  [52] "analysis.metadata.read_groups.library_name"                               
#>  [53] "analysis.metadata.read_groups.library_preparation_kit_catalog_number"     
#>  [54] "analysis.metadata.read_groups.library_preparation_kit_name"               
#>  [55] "analysis.metadata.read_groups.library_preparation_kit_vendor"             
#>  [56] "analysis.metadata.read_groups.library_preparation_kit_version"            
#>  [57] "analysis.metadata.read_groups.library_selection"                          
#>  [58] "analysis.metadata.read_groups.library_strand"                             
#>  [59] "analysis.metadata.read_groups.library_strategy"                           
#>  [60] "analysis.metadata.read_groups.multiplex_barcode"                          
#>  [61] "analysis.metadata.read_groups.platform"                                   
#>  [62] "analysis.metadata.read_groups.read_group_id"                              
#>  [63] "analysis.metadata.read_groups.read_group_name"                            
#>  [64] "analysis.metadata.read_groups.read_group_qcs.adapter_content"             
#>  [65] "analysis.metadata.read_groups.read_group_qcs.basic_statistics"            
#>  [66] "analysis.metadata.read_groups.read_group_qcs.created_datetime"            
#>  [67] "analysis.metadata.read_groups.read_group_qcs.encoding"                    
#>  [68] "analysis.metadata.read_groups.read_group_qcs.fastq_name"                  
#>  [69] "analysis.metadata.read_groups.read_group_qcs.kmer_content"                
#>  [70] "analysis.metadata.read_groups.read_group_qcs.overrepresented_sequences"   
#>  [71] "analysis.metadata.read_groups.read_group_qcs.per_base_n_content"          
#>  [72] "analysis.metadata.read_groups.read_group_qcs.per_base_sequence_content"   
#>  [73] "analysis.metadata.read_groups.read_group_qcs.per_base_sequence_quality"   
#>  [74] "analysis.metadata.read_groups.read_group_qcs.per_sequence_gc_content"     
#>  [75] "analysis.metadata.read_groups.read_group_qcs.per_sequence_quality_score"  
#>  [76] "analysis.metadata.read_groups.read_group_qcs.per_tile_sequence_quality"   
#>  [77] "analysis.metadata.read_groups.read_group_qcs.percent_gc_content"          
#>  [78] "analysis.metadata.read_groups.read_group_qcs.read_group_qc_id"            
#>  [79] "analysis.metadata.read_groups.read_group_qcs.sequence_duplication_levels" 
#>  [80] "analysis.metadata.read_groups.read_group_qcs.sequence_length_distribution"
#>  [81] "analysis.metadata.read_groups.read_group_qcs.state"                       
#>  [82] "analysis.metadata.read_groups.read_group_qcs.submitter_id"                
#>  [83] "analysis.metadata.read_groups.read_group_qcs.total_sequences"             
#>  [84] "analysis.metadata.read_groups.read_group_qcs.updated_datetime"            
#>  [85] "analysis.metadata.read_groups.read_group_qcs.workflow_end_datetime"       
#>  [86] "analysis.metadata.read_groups.read_group_qcs.workflow_link"               
#>  [87] "analysis.metadata.read_groups.read_group_qcs.workflow_start_datetime"     
#>  [88] "analysis.metadata.read_groups.read_group_qcs.workflow_type"               
#>  [89] "analysis.metadata.read_groups.read_group_qcs.workflow_version"            
#>  [90] "analysis.metadata.read_groups.read_length"                                
#>  [91] "analysis.metadata.read_groups.RIN"                                        
#>  [92] "analysis.metadata.read_groups.sequencing_center"                          
#>  [93] "analysis.metadata.read_groups.sequencing_date"                            
#>  [94] "analysis.metadata.read_groups.size_selection_range"                       
#>  [95] "analysis.metadata.read_groups.spike_ins_concentration"                    
#>  [96] "analysis.metadata.read_groups.spike_ins_fasta"                            
#>  [97] "analysis.metadata.read_groups.state"                                      
#>  [98] "analysis.metadata.read_groups.submitter_id"                               
#>  [99] "analysis.metadata.read_groups.target_capture_kit"                         
#> [100] "analysis.metadata.read_groups.target_capture_kit_catalog_number"          
#> [101] "analysis.metadata.read_groups.target_capture_kit_name"                    
#> [102] "analysis.metadata.read_groups.target_capture_kit_target_region"           
#> [103] "analysis.metadata.read_groups.target_capture_kit_vendor"                  
#> [104] "analysis.metadata.read_groups.target_capture_kit_version"                 
#> [105] "analysis.metadata.read_groups.to_trim_adapter_sequence"                   
#> [106] "analysis.metadata.read_groups.updated_datetime"                           
#> [107] "analysis.state"                                                           
#> [108] "analysis.submitter_id"                                                    
#> [109] "analysis.updated_datetime"                                                
#> [110] "analysis.workflow_end_datetime"                                           
#> [111] "analysis.workflow_link"                                                   
#> [112] "analysis.workflow_start_datetime"                                         
#> [113] "analysis.workflow_type"                                                   
#> [114] "analysis.workflow_version"                                                
#> [115] "downstream_analyses.analysis_id"                                          
#> [116] "downstream_analyses.analysis_type"