Prepare GDC manifest file for bulk download

The manifest function/method creates a manifest of files to be downloaded using the GDC Data Transfer Tool. There are methods for creating manifest data frames from GDCQuery objects that contain file information ("cases" and "files" queries).

Usage

manifest(x, from = 0, size = count(x), ...)

# S3 method for gdc_files
manifest(x, from = 0, size = count(x), ...)

# S3 method for GDCfilesResponse
manifest(x, from = 0, size = count(x), ...)

# S3 method for GDCcasesResponse
manifest(x, from = 0, size = count(x), ...)

Arguments

x

An GDCQuery object of subclass "gdc_files" or "gdc_cases".

from

Record number from which to start when returning the manifest.

size

The total number of records to return. Default will return the usually desirable full set of records.

...

passed to PUT.

Value

A tibble, also of type "gdc_manifest", with five columns:

id

filename

md5

size

state

Methods (by class)

manifest(gdc_files):

manifest(GDCfilesResponse):

manifest(GDCcasesResponse):

Examples

gFiles = files()
shortManifest = gFiles %>% manifest(size=10)
head(shortManifest,n=3)
#>                                     id data_format     access
#> 1 1294fc89-fae1-4d71-b970-dee321eeb6c1         VCF controlled
#> 2 e1e7a77e-9d63-4247-b8f3-e2330215e85c         VCF controlled
#> 3 6c59c1f9-8f35-41db-958b-a06047b24d01         VCF controlled
#>                                                                                        file_name
#> 1                      9160d8a8-006d-4fd5-8c9e-09ec6179369a.wxs.muse.raw_somatic_mutation.vcf.gz
#> 2                    e32336ff-87c2-4d5d-9366-ac835157d08d.wxs.pindel.raw_somatic_mutation.vcf.gz
#> 3 MMRF_COMMPASS.124d4920-7fe1-491d-93da-ad76f9eb8f75.wxs.SomaticSniper.somatic_annotation.vcf.gz
#>                           submitter_id               data_category       acl
#> 1 8c7fcd35-5fda-4eff-b368-f373fefc8327 Simple Nucleotide Variation phs000748
#> 2 40ac1eea-f152-403e-9a58-66dd15384d14 Simple Nucleotide Variation phs000748
#> 3 43f6c8d5-835d-47ec-a7c8-22ac6a88c73b Simple Nucleotide Variation phs000748
#>                         type file_size state_comment
#> 1    simple_somatic_mutation     29042            NA
#> 2    simple_somatic_mutation     47937            NA
#> 3 annotated_somatic_mutation   2175870            NA
#>                   created_datetime                           md5sum
#> 1 2019-01-28T10:21:43.275213-06:00 0138a8506fdbca4cf5b6c9e71c61f0b7
#> 2 2019-07-15T10:23:04.984044-05:00 6a94c60d2483df6264f876bee28aa611
#> 3 2022-01-26T21:35:47.361393-06:00 3fff8d28dfde7fb7140b1c25e84b8f6e
#>                   updated_datetime                              file_id
#> 1 2019-06-24T08:07:19.797044-05:00 1294fc89-fae1-4d71-b970-dee321eeb6c1
#> 2 2019-08-21T12:47:37.999949-05:00 e1e7a77e-9d63-4247-b8f3-e2330215e85c
#> 3 2022-01-30T17:20:50.104776-06:00 6c59c1f9-8f35-41db-958b-a06047b24d01
#>   error_type                   data_type    state experimental_strategy
#> 1         NA Raw Simple Somatic Mutation released                   WXS
#> 2         NA Raw Simple Somatic Mutation released                   WXS
#> 3         NA  Annotated Somatic Mutation released                   WXS
#>   proportion_reads_mapped proportion_base_mismatch contamination_error platform
#> 1                      NA                       NA                  NA     <NA>
#> 2                      NA                       NA                  NA     <NA>
#> 3                      NA                       NA                  NA     <NA>
#>   pairs_on_diff_chr total_reads mean_coverage proportion_reads_duplicated
#> 1                NA          NA            NA                          NA
#> 2                NA          NA            NA                          NA
#> 3                NA          NA            NA                          NA
#>   contamination average_base_quality average_insert_size average_read_length
#> 1            NA                   NA                  NA                  NA
#> 2            NA                   NA                  NA                  NA
#> 3            NA                   NA                  NA                  NA
#>   proportion_targets_no_coverage
#> 1                             NA
#> 2                             NA
#> 3                             NA