GDC manifest file creation and file download

manifest() creates a manifest of files to be downloaded using the GDC Data Transfer Tool.

transfer() retrieves files from the manifest using the GDC Data Transfer Tool.

transfer_help() queries the the command line GDC Data Transfer Tool, gdc-client, for available options to be used in the transfer() command.

manifest(uuids, destination_dir = tempfile(), token = NULL)

transfer(manifest, destination_dir = tempfile(), args = character(),
  token_file = NULL, gdc_client = "gdc-client")

transfer_help(gdc_client = "gdc-client")

Arguments

uuids

character() of GDC file UUIDs.

destination_dir

character(1) file path to a directory for downloading the manifest or files. The manifest file name, of the form ‘gdc_manifest_20160610_12345.txt’, must not exist.

token

(optional) character(1) security token allowing access to restricted data. See https://docs.gdc.cancer.gov/API/Users_Guide/Getting_Started/#authentication.

manifest

character(1) file path to manifest created by manifest()

args

character() vector specifying command-line arguments to be passed to gdc-client. See transfer_help for possible values. The arguments --manifest, --dir, and --token-file are determined by manifest, destination_dir, and token, respectively, and should NOT be provided as elements of args.

token_file

character(1) path to a file containing security token allowing access to restricted data. See https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/.

gdc_client

character(1) name or path to gdc-client executable. On Windows, use / or \\ as the file separator.

Value

character(1) directory path to which the files were downloaded.

Examples

uuids <- c("e3228020-1c54-4521-9182-1ea14c5dc0f7",
           "18e1e38e-0f0a-4a0e-918f-08e6201ea140")
(manifest <- manifest(uuids))

#> Error in curl::curl_fetch_disk(url, x$path, handle = handle): SSL connect error

destination <- transfer(manifest)

#> Error in file.exists(manifest): invalid 'file' argument

dir(destination)

#> Error in dir(destination): object 'destination' not found