Prepare GDC manifest file for bulk download

The manifest function/method creates a manifest of files to be downloaded using the GDC Data Transfer Tool. There are methods for creating manifest data frames from GDCQuery objects that contain file information ("cases" and "files" queries).

Usage

manifest(x, from = 0, size = count(x), ...)

# S3 method for gdc_files
manifest(x, from = 0, size = count(x), ...)

# S3 method for GDCfilesResponse
manifest(x, from = 0, size = count(x), ...)

# S3 method for GDCcasesResponse
manifest(x, from = 0, size = count(x), ...)

Arguments

x

An GDCQuery object of subclass "gdc_files" or "gdc_cases".

from

Record number from which to start when returning the manifest.

size

The total number of records to return. Default will return the usually desirable full set of records.

...

passed to PUT.

Value

A tibble, also of type "gdc_manifest", with five columns:

id

filename

md5

size

state

Methods (by class)

manifest(gdc_files):

manifest(GDCfilesResponse):

manifest(GDCcasesResponse):

Examples

gFiles = files()
shortManifest = gFiles %>% manifest(size=10)
head(shortManifest,n=3)
#>                                     id data_format     access
#> 1 c786e2e9-04ea-48cf-b8a5-3daa96c1837b         BAM controlled
#> 2 4f27d16f-0029-43f5-8627-7f477b4cce6d         TSV controlled
#> 3 c7546d88-1748-4760-b06f-0b510484345a         VCF controlled
#>                                                                      file_name
#> 1     c7522bbe-48b8-4338-b36e-564021353f00.rna_seq.transcriptome.gdc_realn.bam
#> 2    c7522bbe-48b8-4338-b36e-564021353f00.rna_seq.star_splice_junctions.tsv.gz
#> 3 bce47c3d-a414-49c9-bd27-9e941144860c.wgs.CaVEMan.raw_somatic_mutation.vcf.gz
#>                           submitter_id               data_category       acl
#> 1 83385997-2e30-4a0e-ba08-8a359cf9e758            Sequencing Reads phs001486
#> 2 b1a12732-5668-4c9a-90b2-ed8817ffa063     Transcriptome Profiling phs001486
#> 3 4d86c0a2-deb6-4060-9a64-1ad96ef94a51 Simple Nucleotide Variation phs001486
#>                      type platform   file_size                 created_datetime
#> 1           aligned_reads Illumina 16264880178 2022-01-05T13:58:13.190285-06:00
#> 2         gene_expression     <NA>     2744657 2022-01-05T16:06:14.065970-06:00
#> 3 simple_somatic_mutation     <NA>     3126634 2020-07-10T11:39:51.205109-05:00
#>                             md5sum                 updated_datetime
#> 1 c1eba4660cf44f9018fe9323bb4c88ec 2022-01-19T23:25:37.395824-06:00
#> 2 7d52913e765f757485357bb7c7bedd19 2022-01-19T23:22:31.836567-06:00
#> 3 23fa4cbce43707f7da9100815d699262 2020-10-26T12:35:23.058290-05:00
#>                                file_id                      data_type    state
#> 1 c786e2e9-04ea-48cf-b8a5-3daa96c1837b                  Aligned Reads released
#> 2 4f27d16f-0029-43f5-8627-7f477b4cce6d Splice Junction Quantification released
#> 3 c7546d88-1748-4760-b06f-0b510484345a    Raw Simple Somatic Mutation released
#>   experimental_strategy state_comment error_type proportion_reads_mapped
#> 1               RNA-Seq            NA         NA                      NA
#> 2               RNA-Seq            NA         NA                      NA
#> 3                   WGS            NA         NA                      NA
#>   proportion_base_mismatch contamination_error pairs_on_diff_chr total_reads
#> 1                       NA                  NA                NA          NA
#> 2                       NA                  NA                NA          NA
#> 3                       NA                  NA                NA          NA
#>   mean_coverage proportion_reads_duplicated contamination average_base_quality
#> 1            NA                          NA            NA                   NA
#> 2            NA                          NA            NA                   NA
#> 3            NA                          NA            NA                   NA
#>   average_insert_size average_read_length proportion_targets_no_coverage
#> 1                  NA                  NA                             NA
#> 2                  NA                  NA                             NA
#> 3                  NA                  NA                             NA