Start a query of GDC metadata

The basis for all functionality in this package starts with constructing a query in R. The GDCQuery object contains the filters, facets, and other parameters that define the returned results. A token is required for accessing certain datasets.

query(entity, token = NULL, filters = NULL, facets = NULL,
  fields = default_fields(entity))

cases(...)

files(...)

projects(...)

annotations(...)

Arguments

entity

character vector of 'cases','files','annotations', or 'projects'

token

a character string reprenting the token

filters

a filter list

facets

a facets list

fields

a list of fields to return

...

passed through to query

Value

An S3 object, the GDCQuery object. This is a list with filters, facets, and fields as members.

filters

facets

fields

token

Functions

cases: convenience contructor for a GDCQuery for cases

files: convenience contructor for a GDCQuery for cases

projects: convenience contructor for a GDCQuery for cases

annotations: convenience contructor for a GDCQuery for annotations