Query GDC for data slices

This function returns a BAM file representing reads overlapping regions specified either as chromosomal regions or as gencode gene symbols.

slicing(uuid, regions, symbols, destination = tempfile(), overwrite = FALSE,
  progress = interactive(), token = NULL, legacy = FALSE)

Arguments

uuid
character(1) identifying the BAM file resource

regions
character() vector describing chromosomal regions, e.g., c("chr1", "chr2:10000", "chr3:10000-20000") (all of chromosome 1, chromosome 2 from position 10000 to the end, chromosome 3 from 10000 to 20000).

symbols
character() vector of gencode gene symbols, e.g., c("BRCA1", "PTEN")

destination
character(1) default tempfile() file path for BAM file slice

overwrite
logical(1) default FALSE can destination be overwritten?

progress
logical(1) default interactive() should a progress bar be used?

token
character(1) security token allowing access to restricted data. Almost all BAM data is restricted, so a token is usually required. See https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/.

legacy
logical(1) whether or not to use the "legacy" archive, containing older, non-harmonized data.

Value

character(1) destination to the downloaded BAM file

Examples

slicing("df80679e-c4d3-487b-934c-fcc782e5d46e",
        regions="chr17:75000000-76000000",
        token=gdc_token())

#> Error in .gdc_post(endpoint = sprintf("legacy/slicing/view/%s", uuid),     add_headers(`Content-type` = "application/json"), write_disk(destination,         overwrite), if (progress) progress() else NULL, body = toJSON(body),     token = token, legacy = legacy): Not Found (HTTP 404).