Bulk data download

The GDC maintains a special tool, https://docs.gdc.cancer.gov/Data_Transfer_Tool/Users_Guide/Getting_Started/, that enables high-performance, potentially parallel, and resumable downloads. The Data Transfer Tool is an external program that requires separate download. Due to recent changes in the GDC API, the transfer function now validates the version of the `gdc-client` to ensure reliable downloads.

Usage

transfer(uuids, args = character(), token = NULL, overwrite = FALSE)

gdc_client_version_validate(valid_version = .GDC_COMPATIBLE_VERSION)

transfer_help()

Arguments

uuids

character() vector of GDC file UUIDs

args

character() vector specifying command-line arguments to be passed to gdc-client. See transfer_help for possible values. The arguments --manifest, --dir, and --token-file are determined by manifest, destination_dir, and token, respectively, and should NOT be provided as elements of args.

token

character(1) containing security token allowing access to restricted data. See https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/. Note that the GDC transfer tool requires a file for data transfer. Therefore, this token will be written to a temporary file (with appropriate permissions set).

overwrite

logical(1) default FALSE indicating whether existing files with identical name should be over-written.

valid_version

character(1) The last known version that works for the current data release for which to validate against, not typically changed by the end-user.

Value

character(1) directory path to which the files were downloaded.

Functions

gdc_client_version_validate(): If you are using the 'client' option, your `gdc-client` should be up-to-date (>= 1.3.0).

transfer_help():

Examples

if (FALSE) {
uuids = files() %>% 
  filter(access == "open") %>% 
  results() %>%
  ids()
file_paths <- transfer(uuids)
file_paths
names(file_paths)
# and with authenication
# REQUIRES gdc_token 
# destination <- transfer(uuids,token=gdc_token())
}